refactor: separate download orchestration and pipeline
